# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ivmp-service API",
        "title": "ivmp-service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/swagger/ivmp-service",
    "paths": {
...
        "/camera/import-cameras": {
            "post": {
                "description": "导入摄像头",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Camera"
                ],
                "summary": "导入摄像头",
                "parameters": [
                    {
                        "description": "csv文件字符串",
                        "name": "csv_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":0,\"data\":{},\"msg\":\"success\"}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}